删除段信息
git-svn-id: svn://127.0.0.1@698 3f3bbce5-bc85-8142-9eb1-77d090cfebd4

+++ main/modem/wms/wms/src/wmse/wmscfg.c
@@ -2124,7 +2124,7 @@ void wms_cfg_delete_dup_info_cache
 )
 {
 #ifdef FEATURE_CDSMS_DUP_DETECTION
-  uint8 i;
+  uint32 i;
 
   /* Find the entry to delete from the cache */
   for(i=0; i<WMS_MESSAGE_LIST_MAX; i++)
